refactor: project 폴더 구분
