+++ src/main/java/com/sejong/elasticservice/project/repository/ProjectElasticDocumentRepository.java
@@ -1,6 +1,7 @@
-package com.sejong.elasticservice.project;
+package com.sejong.elasticservice.project.repository;
 
 
+import com.sejong.elasticservice.project.domain.ProjectDocument;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 public interface ProjectElasticDocumentRepository extends ElasticsearchRepository<ProjectDocument,String> {
